Roadmap 1.1 DLL layout self-check and 1.5 warning/bug fixes
1.1: LibSm64::layoutCheckReport() reads game state through the copied
decomp structs and cross-checks relationships the game guarantees: Mario's
object sits in gObjectPool at a multiple of sizeof(Object), its behavior
is bhvMario, MarioState::marioObj points back, oPos and header.gfx.pos
mirror MarioState::pos, the floor normal is unit length, pointers land in
.data/.bss; in lightweight mode, every hot symbol must lie inside the
saved slices (now a shared constexpr table used by save, load and the
check). Resource::verifyLayout() throws on failure and runs once per
scattershot thread after the start frame loads. tasfw-tools/dllcheck runs
it standalone and prints frame-advance and save/load cost. Both the pinned
2022 DLL and wafel's 2023 DLL pass every check; first Tier B numbers:
frame advance ~10 us, lightweight save/load 285/42 us, full 1400/190 us.

1.5: fix the TurnAround lambda that could fall off the end (C4715), the
always-false && in TurnUphill_1f, two empty-body if(...); statements in
the Approach/Recover stages (now return false), eight discarded .executed
results, printf %d with 64-bit values, the root Segment argument order
(RngHash truncated into nScripts), Rotation::Negate, unhandled switch
cases, missing overrides, main returning false, a bool/s32 compare in
PyramidUpdate and int16->int8 narrowing in Inputs.cpp. Both compilers now
build with only third-party warnings (plus three MSVC C4244 in
TiltTargetShot initializer lists). TASFW_WARNINGS_AS_ERRORS option added,
off by default.

Also: clang-cl was silently ignoring -ffp-contract=off (only a
-Wunknown-argument warning), so FP contraction had stayed on; it is now
passed as /clang:-ffp-contract=off and the clang baseline re-measured.
CI workflow now runs on every push. Perf gate: 0 regressions on MSVC and
clang.

+++ tasfw-scattershot/inc/Scattershot.hpp
@@ -285,7 +285,7 @@ private:
                 }
             });
 
-        printf("Found %d solutions in %d shots.\n", scattershot.Solutions.size(), scattershot.TotalShots);
+        printf("Found %llu solutions in %llu shots.\n", (unsigned long long)scattershot.Solutions.size(), (unsigned long long)scattershot.TotalShots);
 
         std::vector<ScattershotSolution<TOutputState>> solutions;
         solutions.reserve(scattershot.Solutions.size());
